CEREZA manual completa integrada a totales

+++ flujo/views_backup_cereza.py
@@ -286,7 +286,7 @@ def detalle_movimientos_real(request):
                 "cpbnum": m.cpbnum or "",
                 "cuenta": m.cuenta_banco.codigo if m.cuenta_banco else "",
                 "descripcion": m.descripcion or "",
-                "monto": int(monto),
+                "monto": float(monto),
             })
 
         return JsonResponse({
@@ -845,4 +845,3 @@ def inicio(request):
         "diferencia_tc_concepto_id": diferencia_tc_concepto_id,
         "fila_diferencia_tc": fila_diferencia_tc,
     })
-
